+++ modules/gitrepo/compare_test.go
@@ -7,6 +7,7 @@ import (
 	"testing"
 
 	"code.gitea.io/gitea/modules/git"
+
 	"github.com/stretchr/testify/assert"
 )
